+++ lib/components/article_sentences.dart
@@ -9,22 +9,26 @@ import '../models/article_titles.dart';
 import '../models/settings.dart';
 import '../models/word.dart';
 import '../models/sentence.dart';
+import '../models/global.dart';
+import '../store/store.dart';
+import '../store/wordwise.dart';
 
-import './article_richtext.dart';
+import './article_sentence.dart';
 import '../functions/article.dart';
 import '../functions/utility.dart';
 
 // init text style
-TextStyle bodyTextStyle = TextStyle(
-    color: Colors.black87, fontSize: 20.0, fontFamily: 'NotoSans-Medium');
+TextStyle bodyTextStyle = TextStyle(color: Colors.black87, fontSize: 20.0, fontFamily: 'NotoSans-Medium');
 
 class ArticleSentences extends StatefulWidget {
   ArticleSentences(
       {Key key,
       @required this.article,
       @required this.sentences,
+      this.needWordWise = true,
       this.crossAxisAlignment = CrossAxisAlignment.start})
       : super(key: key);
+  final bool needWordWise;
   final Article article; // 用于计算文章的未掌握数目
   final List<Sentence> sentences; // 需要渲染的多条句子
   final CrossAxisAlignment crossAxisAlignment; // 句子的位置
@@ -34,6 +38,7 @@ class ArticleSentences extends StatefulWidget {
 }
 
 class ArticleSentencesState extends State<ArticleSentences> {
+  Global _global;
   Map seekTextSpanTapStatus = Map<String, bool>();
 
   RegExp _startExp = RegExp(r"00[0-9]+\.[0-9]+00");
@@ -41,7 +46,7 @@ class ArticleSentencesState extends State<ArticleSentences> {
   // 后台返回的文章结构
   String _tapedText = ''; // 当前点击的文本
   String _lastTapedText = ''; // 上次点击的文本
-  Settings settings;
+  SettingNews _settings;
 
   // 必学的高亮色
   TextStyle needLearnTextStyle;
@@ -51,14 +56,13 @@ class ArticleSentencesState extends State<ArticleSentences> {
   @override
   initState() {
     super.initState();
-    settings = Provider.of<Settings>(context, listen: false);
+    _settings = Provider.of<SettingNews>(context, listen: false);
+    _global = Provider.of<Global>(context, listen: false);
   }
 
   int _getIDByTitle(String title) {
     var articles = Provider.of<ArticleTitles>(context, listen: false);
-    var titles = articles.titles
-        .where((d) => d.title.toLowerCase() == title.toLowerCase())
-        .toList();
+    var titles = articles.titles.where((d) => d.title.toLowerCase() == title.toLowerCase()).toList();
     if (titles.length > 0) {
       return titles[0].id;
     }
@@ -80,11 +84,18 @@ class ArticleSentencesState extends State<ArticleSentences> {
         setState(() {
           word.learned = learned;
         });
+        //update global word status
+        _global.words[word.text.toLowerCase()] = word;
+
+        await getDefinitionByWord(word.text.toLowerCase());
+        setState(() {});
+
         // set all sentences word to this state
         await widget.article.putLearned(word); //重新计算文章未掌握单词数
         var articleTitles = Provider.of<ArticleTitles>(context, listen: false);
-        articleTitles.setUnlearnedCountByArticleID(
-            widget.article.unlearnedCount, widget.article.articleID);
+        articleTitles.setUnlearnedCountByArticleID(widget.article.unlearnedCount, widget.article.articleID);
+        //save to local
+        widget.article.updateLocal();
       }
       ..onTap = (i) {
         // 避免长按的同时触发
@@ -99,8 +110,7 @@ class ArticleSentencesState extends State<ArticleSentences> {
           });
           // 无需学的, 没必要显示级别
           if (word.level != 0)
-            Alert.toast(context, word.level.toString(),
-                position: ToastPosition.bottom, duration: ToastDuration.long);
+            Alert.toast(context, word.level.toString(), position: ToastPosition.bottom, duration: ToastDuration.long);
           // 实时增加次数的效果
           widget.article.increaseLearnCount(word.text);
           // 记录学习次数
@@ -110,7 +120,7 @@ class ArticleSentencesState extends State<ArticleSentences> {
           // 已经在这个单词页, 就不要跳转了
           if (_lastTapedText.toLowerCase() == word.text.toLowerCase() &&
               word.text.toLowerCase() != widget.article.title.toLowerCase() &&
-              settings.isJump) {
+              _settings.isJump) {
             int id = _getIDByTitle(word.text);
             if (id != 0) {
               Navigator.pushNamed(context, '/Article', arguments: id);
@@ -124,8 +134,7 @@ class ArticleSentencesState extends State<ArticleSentences> {
 
   // 生成修改播放位置的图标
   TextSpan getSeekTextSpan(BuildContext context, String time, Sentence s) {
-    if (seekTextSpanTapStatus[time] == null)
-      seekTextSpanTapStatus[time] = false;
+    if (seekTextSpanTapStatus[time] == null) seekTextSpanTapStatus[time] = false;
     Duration seekTime = toDuration(time);
     TapGestureRecognizer recognizer = TapGestureRecognizer()
       ..onTap = () {
@@ -139,13 +148,10 @@ class ArticleSentencesState extends State<ArticleSentences> {
           });
         });
       };
-    TextStyle playTextStyle =
-        bodyTextStyle.copyWith(color: Theme.of(context).primaryColorLight);
+    TextStyle playTextStyle = bodyTextStyle.copyWith(color: Theme.of(context).primaryColorLight);
     return TextSpan(
         text: " ▷ ",
-        style: seekTextSpanTapStatus[time] || s.highlight
-            ? playTextStyle.copyWith(fontWeight: FontWeight.bold)
-            : playTextStyle,
+        style: seekTextSpanTapStatus[time] || s.highlight ? playTextStyle.copyWith(fontWeight: FontWeight.bold) : playTextStyle,
         recognizer: recognizer);
   }
 
@@ -157,20 +163,16 @@ class ArticleSentencesState extends State<ArticleSentences> {
   // 定义应该的 style
   TextStyle _defineStyle(Word word) {
     bool isCommandWord = (word.level != null && word.level != 0); // 是否3000常用
-    bool isSelected =
-        (_tapedText.toLowerCase() == word.text.toLowerCase()); // 是否选中
+    bool isSelected = (_tapedText.toLowerCase() == word.text.toLowerCase()); // 是否选中
     // 常用高亮色
-    this.needLearnTextStyle =
-        bodyTextStyle.copyWith(color: Theme.of(context).primaryColorLight);
+    this.needLearnTextStyle = bodyTextStyle.copyWith(color: Theme.of(context).primaryColorLight);
     // 非常用的高亮色
-    this.noNeedLearnTextStyle =
-        bodyTextStyle.copyWith(color: Theme.of(context).primaryColorDark);
+    this.noNeedLearnTextStyle = bodyTextStyle.copyWith(color: Theme.of(context).primaryColorDark);
 
     //根据条件逐步加工修改的样式
     TextStyle processTextStyle = bodyTextStyle;
     // 是常用
-    processTextStyle =
-        isCommandWord ? needLearnTextStyle : noNeedLearnTextStyle;
+    processTextStyle = isCommandWord ? needLearnTextStyle : noNeedLearnTextStyle;
     //无需学习的非标准单词
     if (!isNeedLearn(word)) processTextStyle = bodyTextStyle;
     // 已经学会且没有选中, 不用任何修改
@@ -184,14 +186,18 @@ class ArticleSentencesState extends State<ArticleSentences> {
     // 查找时高亮
     if (widget.article.findWord.toLowerCase() == word.text.toLowerCase())
       processTextStyle = processTextStyle.copyWith(
-          color: Colors.deepOrange[700],
-          decoration: TextDecoration.underline,
-          decorationStyle: TextDecorationStyle.wavy);
+          color: Colors.deepOrange[700], decoration: TextDecoration.underline, decorationStyle: TextDecorationStyle.wavy);
     return processTextStyle;
+  }
+
+  keepWordHasSameStat(Word word) {
+    if (_global.words.containsKey(word.text.toLowerCase())) word.learned = _global.words[word.text.toLowerCase()].learned;
+    return word;
   }
 
 // 组装为需要的 textSpan
   TextSpan getTextSpan(Word word) {
+    word = keepWordHasSameStat(word);
     if (word.text == "\n" || _startExp.hasMatch(word.text)) {
       return TextSpan(text: "");
     }
@@ -200,23 +206,20 @@ class ArticleSentencesState extends State<ArticleSentences> {
     TextSpan subscript = TextSpan(); // 显示该单词查询次数的下标
 
     //需要学习的单词
-    if (word.learned == false && hasLetter(word.text) && word.text.length > 1) {
-      //有查询下标则显示
-      if (word.count != 0) {
-        subscript = TextSpan(
-            text: word.count.toString(),
-            style: wordStyle.copyWith(fontSize: 12)); // 数字样式和原本保持一致, 只是变小
-      }
+    if (word.learned == false && isNeedLearn(word)) {
+      String shortDef = "";
+      if (widget.needWordWise && Store.wordwiseMap[word.text.toLowerCase()] != null)
+        shortDef = "-" + Store.wordwiseMap[word.text.toLowerCase()];
+      if (word.count != 0) shortDef += " " + word.count.toString();
+      subscript = TextSpan(text: shortDef, style: wordStyle.copyWith(fontSize: 12)); // 下标样式和原本保持一致, 只是变小
+
     }
 
     return TextSpan(text: _getBlank(word.text), children: [
       // if not letter no need recognizer
       hasLetter(word.text)
-          ? TextSpan(
-              text: word.text,
-              style: wordStyle,
-              recognizer: _getTapRecognizer(word))
-          : TextSpan(text: word.text, style: wordStyle),
+          ? TextSpan(text: word.text, style: wordStyle, recognizer: _getTapRecognizer(word))
+          : TextSpan(text: word.text, style: wordStyle, recognizer: _getTapRecognizer(word)),
       subscript,
     ]);
   }
@@ -232,20 +235,21 @@ class ArticleSentencesState extends State<ArticleSentences> {
     return star;
   }
 
-  ArticleRichText buildArticleRichText(Sentence s) {
+  ArticleSentence buildArticleRichText(Sentence s) {
     TextSpan star = getSeekButton(context, s.words[0].text, s);
     List<TextSpan> words = s.words.map((d) {
       return getTextSpan(d);
     }).toList();
     words.insert(0, star);
+
     TextStyle playingStyle = TextStyle();
     //if play to current sentence
-    if (s.highlight) {
+    if (s.highlight)
       playingStyle = TextStyle(
         backgroundColor: Colors.teal[50],
       );
-    }
-    return ArticleRichText(
+
+    return ArticleSentence(
         textSpan: TextSpan(
           style: playingStyle,
           text: "",
@@ -257,13 +261,11 @@ class ArticleSentencesState extends State<ArticleSentences> {
   @override
   Widget build(BuildContext context) {
     //只有一个单词时候不要用 Column封装,避免位置上移
-    if (widget.sentences.length == 1)
-      return buildArticleRichText(widget.sentences[0]);
+    if (widget.sentences.length == 1) return buildArticleRichText(widget.sentences[0]);
     List<Widget> richTextList = widget.sentences.map((s) {
       return buildArticleRichText(s);
     }).toList();
 
-    return Column(
-        children: richTextList, crossAxisAlignment: CrossAxisAlignment.start);
+    return Column(children: richTextList, crossAxisAlignment: CrossAxisAlignment.start);
   }
 }