+++ lib/components/config_autoplay.dart
@@ -8,14 +8,15 @@ import '../models/settings.dart';
 class ConfigAutoPlay extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Consumer<Settings>(builder: (context, setting, child) {
-      return Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-        Text('Autoplay'),
-        Switch(
-          value: setting.isAutoplay,
-          onChanged: setting.setIsAutoplay,
-        )
-      ]);
-    });
+    return Selector<SettingNews, bool>(
+      selector: (context, settings) => settings.isAutoplay,
+      builder: (context, isAutoplay, child) {
+        return SwitchListTile(
+            value: isAutoplay, onChanged: Provider.of<SettingNews>(context, listen: false).setIsAutoplay, title: child);
+      },
+      child: Text(
+        'Autoplay',
+      ),
+    );
   }
 }