+++ lib/components/config_dark_theme.dart
@@ -8,14 +8,15 @@ import '../models/settings.dart';
 class ConfigDarkTheme extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Consumer<Settings>(builder: (context, setting, child) {
-      return Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-        Text('Dark Mode'),
-        Switch(
-          value: setting.isDark,
-          onChanged: setting.setIsDark,
-        )
-      ]);
-    });
+    return Selector<SettingNews, bool>(
+        selector: (context, settings) => settings.isDark,
+        builder: (context, isDark, child) {
+          return SwitchListTile(
+              value: isDark,
+              onChanged: Provider.of<SettingNews>(context, listen: false).setIsDark,
+              title: Text(
+                'Dark Mode',
+              ));
+        });
   }
 }