+++ lib/models/article.dart
@@ -1,9 +1,9 @@
 // 文章详情内容
 import 'dart:async';
 import 'dart:convert';
+//import '../youtube_player_flutter/youtube_player_flutter.dart';
 import 'package:youtube_player_flutter/youtube_player_flutter.dart';
 
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -11,6 +11,7 @@ import '../functions/article.dart';
 import './sentence.dart';
 import './word.dart';
 import '../store/store.dart';
+import '../store/wordwise.dart';
 
 class Article with ChangeNotifier {
   bool checkSentenceHighlight = false;
@@ -52,7 +53,7 @@ class Article with ChangeNotifier {
   }
 
   // 从 json 中设置
-  setFromJSON(Map json) {
+  setFromJSON(Map json) async {
     this.articleID = json['id'];
     this.title = json['title'];
     this.youtube = json['Youtube'];
@@ -65,17 +66,33 @@ class Article with ChangeNotifier {
     this.unlearnedCount = json['UnlearnedCount'];
     this.avatar = json['Avatar'];
     this.wordCount = json['WordCount'];
-    //notifyListeners2();
   }
 
-  /*
-  clear() {
-    this.youtube = '';
-    this.title = '';
-    this.sentences.clear();
-    // notifyListeners2();
+  queryWordWise() async {
+    for (var i = 0; i < this.sentences.length; i++) {
+      for (var j = 0; j < this.sentences[i].words.length; j++) {
+        Word word = this.sentences[i].words[j];
+        if (isNeedLearn(word) && word.learned == false) {
+          await getDefinitionByWord(word.text.toLowerCase());
+        }
+      }
+    }
   }
-   */
+
+  updateLocal() {
+    this.setToLocal(jsonEncode(this));
+  }
+
+  // trans to json string
+  Map<String, dynamic> toJson() => {
+        'id': this.articleID,
+        'title': this.title,
+        'Youtube': this.youtube,
+        'Sentences': jsonEncode(this.sentences),
+        'UnlearnedCount': this.unlearnedCount,
+        'Avatar': this.avatar,
+        'WordCount': this.wordCount,
+      };
 
   setToLocal(String data) async {
     // 登录后存储到临时缓存中
@@ -97,9 +114,8 @@ class Article with ChangeNotifier {
   // justUpdateLocal 仅更新本地缓存, 避免延迟导致页面内容错乱
   Future getArticleByID(int articleID) async {
     this.articleID = articleID;
-    Dio dio = getDio();
-    var response =
-        await dio.get(Store.baseURL + "article/" + this.articleID.toString());
+    var response = await Store.dio()
+        .get(Store.baseURL + "article/" + this.articleID.toString());
 
     this.setFromJSON(response.data);
     this.setToLocal(json.encode(response.data));
@@ -128,8 +144,8 @@ class Article with ChangeNotifier {
     unlearnedCount = allWords.toSet().length;
     unlearnedCount--;
     // 设置本地的列表
-    Dio dio = getDio();
-    var response = await dio.put(Store.baseURL + "article/unlearned_count",
+    var response = await Store.dio().put(
+        Store.baseURL + "article/unlearned_count",
         data: {"article_id": articleID, "unlearned_count": unlearnedCount});
     return response;
   }
@@ -161,6 +177,8 @@ class Article with ChangeNotifier {
   Future putLearned(Word word) async {
     // 标记所有单词为对应状态, 并通知
     this._setWordIsLearned(word.text, word.learned);
-    return word.putLearned().then((d) => _putUnlearnedCount());
+    var d = await word.putLearned();
+    _putUnlearnedCount();
+    return d;
   }
 }