+++ lib/models/article_inherited.dart
@@ -18,7 +18,6 @@ class ArticleInherited extends InheritedWidget {
 
   @override
   bool updateShouldNotify(ArticleInherited old) {
-    // article 有任何变化都会引起通知, 这点很不好
-    return this.article != old.article;
+    return this.article.unlearnedCount != old.article.unlearnedCount;
   }
 }