+++ lib/models/oauth_info.dart
@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import '../store/sign.dart';
+import '../store/store.dart';
+import './controller.dart';
 
 class OauthInfo with ChangeNotifier {
+  Controller controller; // use to show info
   String accessToken;
   String email;
   String name;
   String avatarURL;
+  bool loading = false;
   GoogleSignIn _googleSignIn;
   GoogleSignInAccount _currentUser;
-  Function setAccessTokenCallBack;
+  Future Function() setAccessTokenCallBack;
 
   OauthInfo() {
-    // set callback to _googleSignIn
     _googleSignIn = GoogleSignIn(
       scopes: <String>[
         'profile',
@@ -21,87 +23,80 @@ class OauthInfo with ChangeNotifier {
     );
     _googleSignIn.onCurrentUserChanged.listen((GoogleSignInAccount account) {
       _currentUser = account;
-      if (_currentUser != null) {
-        _handleGetContact();
-      }
-    });
-    _googleSignIn.onCurrentUserChanged.listen((GoogleSignInAccount account) {
-      if (account != null) {
-        account.authentication
-            .then((GoogleSignInAuthentication authentication) {
-          // google 用户注册到服务器后, 记录 token
-          putAccount(account, authentication).then((d) {
-            this.setAccessToken(authentication.accessToken, account.email,
-                account.displayName, account.photoUrl);
-          });
-        });
-      }
+      if (_currentUser != null) handleGetContact();
     });
   }
 
-  Future _handleGetContact() async {
-    _currentUser.authentication
-        .then((GoogleSignInAuthentication authentication) {
-      // google 用户注册到服务器后, 记录 token
-      putAccount(_currentUser, authentication).then((d) {
-        this.setAccessToken(authentication.accessToken, _currentUser.email,
-            _currentUser.displayName, _currentUser.photoUrl);
-      });
+  signoDone() {
+    if (setAccessTokenCallBack != null) setAccessTokenCallBack();
+    this.loading = false;
+    notifyListeners();
+  }
+
+  handleGetContact() async {
+    GoogleSignInAuthentication authentication =
+        await _currentUser.authentication;
+    //put user info to server
+    await putAccount(_currentUser, authentication);
+    this.setToShared(authentication.accessToken, _currentUser.email,
+        _currentUser.displayName, _currentUser.photoUrl);
+    this.signoDone();
+  }
+
+  switchUser() async {
+    await this.disconnect();
+    signIn();
+  }
+
+  Future disconnect() async {
+    return _googleSignIn.disconnect().then((e) {
+      String info = "sign out Error: $e";
+      debugPrint(info);
+      if (this.controller != null) this.controller.showSnackBar(info);
     });
   }
 
-  Future switchUser() async {
-    try {
-      await _googleSignIn.disconnect();
-    } catch (error) {
-      print(error);
-    }
-    return signIn();
+  signIn() async {
+    this.loading = true;
+    notifyListeners();
+    _googleSignIn.signIn().catchError((e) {
+      String info = "Please check your network!\nError: $e \nTry again...";
+      debugPrint(info);
+      if (this.controller != null) this.controller.showSnackBar(info);
+      Future.delayed(Duration(seconds: 4), () => this.signIn());
+    });
   }
 
-  Future signIn() async {
-    print("signIn");
-    return _googleSignIn.signIn();
-  }
-
-  setAccessToken(
+  // set login info to shared
+  setToShared(
       String accessToken, String email, String name, String avatarURL) async {
     // 如果从未登录转换到登录, 那么返回需要跳转
     this.accessToken = accessToken;
     this.email = email;
     this.name = name;
     this.avatarURL = avatarURL;
-    await _setToShared();
-    setAccessTokenCallBack();
-    notifyListeners();
-  }
-
-  Future backFromShared() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    this.email = prefs.getString('email');
-    if (this.email != null) {
-      this.setAccessToken(prefs.getString('accessToken'), this.email,
-          prefs.getString('name'), prefs.getString('avatarURL'));
-      //已经登录就同步列表
-      return setAccessTokenCallBack();
-    } else {
-      return this.signIn();
-    }
-  }
-
-  _setToShared() async {
-    // 登录后存储到临时缓存中
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs
+    await Store.prefs
       ..setString('accessToken', this.accessToken)
       ..setString('email', this.email)
       ..setString('name', this.name)
       ..setString('avatarURL', this.avatarURL);
   }
 
-  _removeShared() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs
+  backFromShared() async {
+    var prefs = await Store.prefs;
+    this.email = prefs.getString('email');
+    this.accessToken = prefs.getString('accessToken');
+    this.name = prefs.getString('name');
+    this.avatarURL = prefs.getString('avatarURL');
+    // if is logined, run callback to get articlelist
+    if (this.accessToken != null)
+      this.signoDone();
+    else
+      this.signIn();
+  }
+
+  removeFromShared() async {
+    await Store.prefs
       ..remove('accessToken')
       ..remove('email')
       ..remove('name')
@@ -109,9 +104,9 @@ class OauthInfo with ChangeNotifier {
   }
 
   signOut() async {
-    await _googleSignIn.disconnect();
+    this.loading = true;
+    await this.disconnect();
     this.email = null;
-    _removeShared();
-    notifyListeners();
+    removeFromShared();
   }
 }