+++ lib/pages/article.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'package:flutter/cupertino.dart';
-
 import 'package:flutter/material.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:provider/provider.dart';
+import 'package:visibility_detector/visibility_detector.dart';
 
 import '../components/article_sentences.dart';
 import '../components/article_top_bar.dart';
@@ -15,6 +15,7 @@ import '../models/article.dart';
 import '../models/settings.dart';
 import '../models/sentence.dart';
 import '../models/article_inherited.dart';
+import '../models/controller.dart';
 import '../functions/utility.dart';
 
 class TimeSentenceIndex {
@@ -37,26 +38,25 @@ class ArticlePage extends StatefulWidget {
   _ArticlePageState createState() => _ArticlePageState();
 }
 
-class _ArticlePageState extends State<ArticlePage>
-    with AutomaticKeepAliveClientMixin {
+class _ArticlePageState extends State<ArticlePage> with AutomaticKeepAliveClientMixin {
   bool wantKeepAlive = true;
   Article _article;
   ScrollController _scrollController;
   ArticleTitles _articleTitles;
-  Settings settings;
+  SettingNews settings;
   int _articleID;
-  bool _loading = false;
+  bool _loading = true;
   Timer _timer;
   int _highlightSentenceIndex;
-  // keep {int timeSecond: [SentenceIndex1, SentenceIndex2, ...]}
   List<TimeSentenceIndex> _timeSentenceIndexs = List();
 
   @override
   void initState() {
     super.initState();
     _articleID = widget._articleID;
+
     _scrollController = ScrollController();
-    settings = Provider.of<Settings>(context, listen: false);
+    settings = Provider.of<SettingNews>(context, listen: false);
     _article = Article();
     _articleTitles = Provider.of<ArticleTitles>(context, listen: false);
     _article.articleID = _articleID;
@@ -64,7 +64,6 @@ class _ArticlePageState extends State<ArticlePage>
     _article.notifyListeners2 = () {
       if (this.mounted) setState(() {});
     };
-    _articleTitles.setInstanceArticles(_article);
     loadArticleByID();
     preload();
   }
@@ -81,7 +80,7 @@ class _ArticlePageState extends State<ArticlePage>
   @override
   void dispose() {
     debugPrint("ArticlePage dispose");
-    //为了避免内存泄露，需要调用_controller.dispose
+    //为了避免内存泄露，需要调用_scrollController.dispose
     _scrollController.dispose();
     _timer?.cancel();
     super.dispose();
@@ -90,23 +89,21 @@ class _ArticlePageState extends State<ArticlePage>
   //star check sentence highlight routine
   initRoutine() {
     if (_timer == null && _article.youtube != "") {
-      _timer = Timer.periodic(const Duration(milliseconds: 800),
-          (t) => routineCheckSentenceHighLight());
+      _timer = Timer.periodic(const Duration(milliseconds: 800), (t) => routineCheckSentenceHighLight());
     }
   }
 
   routineCheckSentenceHighLight() {
+    Controller controller = Provider.of<Controller>(context, listen: false);
     //if leave this article page no need checkSentenceHighlight
     if (!_article.checkSentenceHighlight) return;
-    debugPrint("routineCheckSentenceHighLight article=" +
-        widget._articleID.toString());
+    //debugPrint("routineCheckSentenceHighLight article=" + widget._articleID.toString());
     if (_article.youtubeController == null) return;
     int currentIndex;
     for (int i = 0; i < _timeSentenceIndexs.length; i++) {
       int currentSeconds = _article.youtubeController.value.position.inSeconds;
       // current playing time between start and end then highlight it
-      if (currentSeconds >= _timeSentenceIndexs[i].startSeconds &&
-          currentSeconds < _timeSentenceIndexs[i].endSeconds) {
+      if (currentSeconds >= _timeSentenceIndexs[i].startSeconds && currentSeconds < _timeSentenceIndexs[i].endSeconds) {
         _timeSentenceIndexs[i].setHighlight(true, _article.sentences);
         currentIndex = i;
       } else
@@ -114,17 +111,24 @@ class _ArticlePageState extends State<ArticlePage>
     }
     // trigger setState if set new highlight sentence
     if (currentIndex != null && _highlightSentenceIndex != currentIndex) {
-      print("find some set index=" + currentIndex.toString());
-      setState(() {
-        _highlightSentenceIndex = currentIndex;
-      });
+      //make highlight show
+      setState(() {});
+      //auto scroll sentence to top
+      if (settings.isScrollWithPlay &&
+          controller.homeIndex == ArticlePageViewPageIndex &&
+          controller.selectedArticleID == _article.articleID &&
+          _article.youtubeController.value.isPlaying) {
+        int sentenceIndex = _timeSentenceIndexs[currentIndex].indexs[0];
+        Scrollable.ensureVisible(_article.sentences[sentenceIndex].c, duration: Duration(milliseconds: 1400), alignment: 0.0);
+      }
+      //alignment: 0.0);
+      _highlightSentenceIndex = currentIndex;
     }
   }
 
   // preload last and next article from server save to local
   preload() {
-    List<int> result =
-        _articleTitles.findLastNextArticleByID(_article.articleID);
+    List<int> result = _articleTitles.findLastNextArticleByID(_article.articleID);
     int lastID = result[0];
     int nextID = result[1];
 
@@ -137,30 +141,34 @@ class _ArticlePageState extends State<ArticlePage>
     await _article.getArticleByID(_article.articleID);
     if (this.mounted) {
       // 更新本地未学单词数
-      _articleTitles.setUnlearnedCountByArticleID(
-          _article.unlearnedCount, _article.articleID);
+      _articleTitles.setUnlearnedCountByArticleID(_article.unlearnedCount, _article.articleID);
     }
-    _loading = false;
   }
 
   Future loadArticleByID() async {
-    setState(() {
-      _loading = true;
-    });
     bool hasLocal = await _article.getFromLocal(_article.articleID);
-    if (hasLocal) {
-      //如果缓存取到, 就不要更新页面内容, 避免后置更新导致页面跳变
+    if (hasLocal)
+      loadFromServer();
+    else {
+      if (this.mounted)
+        setState(() {
+          _loading = true;
+        });
+      await loadFromServer();
+    }
+
+    if (_article.youtube != null && _article.youtube != "") {
+      this.splitSentencesByTime();
+      this.initRoutine();
+    }
+    if (this.mounted)
       setState(() {
         _loading = false;
       });
-      // use preload replace loadFromServer even get from local
-      //loadFromServer();
-    } else {
-      await loadFromServer();
-      setState(() {});
-    }
-    this.splitSentencesByTime();
-    this.initRoutine();
+    await _article.queryWordWise();
+    // make sure show word wise
+    setState(() {});
+    return hasLocal;
   }
 
   //split article sentences by time
@@ -170,15 +178,12 @@ class _ArticlePageState extends State<ArticlePage>
       String startTime = _article.sentences[i].startTime;
       if (startTime != "") {
         int iStartTime = toDuration(startTime).inSeconds;
-        if (_timeSentenceIndexs.length > 0)
-          _timeSentenceIndexs[_timeSentenceIndexs.length - 1].endSeconds =
-              iStartTime;
+        if (_timeSentenceIndexs.length > 0) _timeSentenceIndexs[_timeSentenceIndexs.length - 1].endSeconds = iStartTime;
         TimeSentenceIndex timeSentenceIndex = TimeSentenceIndex();
         timeSentenceIndex.startSeconds = iStartTime;
         timeSentenceIndex.indexs.add(i);
 
-        while (i + 1 < _article.sentences.length &&
-            _article.sentences[i + 1].startTime == "") {
+        while (i + 1 < _article.sentences.length && _article.sentences[i + 1].startTime == "") {
           i++;
           timeSentenceIndex.indexs.add(i);
         }
@@ -187,8 +192,7 @@ class _ArticlePageState extends State<ArticlePage>
     }
     //make sure last sentence has endtime
     _timeSentenceIndexs[_timeSentenceIndexs.length - 1].endSeconds =
-        _timeSentenceIndexs[_timeSentenceIndexs.length - 1].startSeconds +
-            10000;
+        _timeSentenceIndexs[_timeSentenceIndexs.length - 1].startSeconds + 10000;
   }
 
   Widget refreshBody() {
@@ -201,7 +205,7 @@ class _ArticlePageState extends State<ArticlePage>
   }
 
   Widget body() {
-    return ModalProgressHUD(
+    ModalProgressHUD hud = ModalProgressHUD(
         opacity: 1,
         progressIndicator: getSpinkitProgressIndicator(context),
         color: Theme.of(context).scaffoldBackgroundColor,
@@ -213,6 +217,12 @@ class _ArticlePageState extends State<ArticlePage>
           refreshBody()
         ]),
         inAsyncCall: _loading);
+    return VisibilityDetector(
+        key: Key(_article.articleID.toString()),
+        onVisibilityChanged: (d) {
+          //if (d.visibleFraction == 1) setState(() {});
+        },
+        child: hud);
   }
 
   Widget articleBody() {
@@ -223,10 +233,7 @@ class _ArticlePageState extends State<ArticlePage>
         child: Column(children: [
           ArticleTopBar(article: _article),
           NotMasteredVocabulary(),
-          Padding(
-              padding: EdgeInsets.all(5),
-              child: ArticleSentences(
-                  article: _article, sentences: _article.sentences)),
+          Padding(padding: EdgeInsets.all(5), child: ArticleSentences(article: _article, sentences: _article.sentences)),
         ]));
   }
 
@@ -243,15 +250,11 @@ class _ArticlePageState extends State<ArticlePage>
   Widget build(BuildContext context) {
     super.build(context);
     if (widget._articleID != -1 && widget._articleID != _article.articleID) {
-      wantKeepAlive = false;
       _article.articleID = widget._articleID;
-      loadArticleByID().then((d) => wantKeepAlive = true);
+      this.loadArticleByID();
     }
 
-    print("build article");
     return ArticleInherited(
-        article: this._article,
-        child: Scaffold(
-            body: body(), floatingActionButton: ArticleFloatingActionButton()));
+        article: this._article, child: Scaffold(body: body(), floatingActionButton: ArticleFloatingActionButton()));
   }
 }