+++ lib/pages/article_page_view.dart
@@ -3,48 +3,46 @@ import 'package:provider/provider.dart';
 import '../models/article_titles.dart';
 import './article.dart';
 import '../models/controller.dart';
+import '../models/article_title.dart';
 
 class ArticlePageViewPage extends StatefulWidget {
   @override
   _ArticlePageViewPage createState() => _ArticlePageViewPage();
 }
 
-class _ArticlePageViewPage extends State<ArticlePageViewPage>
-    with AutomaticKeepAliveClientMixin {
-  @override
-  bool get wantKeepAlive => true;
-  Controller _controller;
-  ArticleTitles _articleTitles;
+class _ArticlePageViewPage extends State<ArticlePageViewPage> {
+  //with AutomaticKeepAliveClientMixin {
+  //@override
+  //bool get wantKeepAlive => false;
   @override
   void initState() {
-    _controller = Provider.of<Controller>(context, listen: false);
-    _articleTitles = Provider.of<ArticleTitles>(context, listen: false);
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    super.build(context);
-    if (_controller.articlePageController == null) return Container();
+    //super.build(context);
     return WillPopScope(
       onWillPop: () async {
-        if (_controller.mainSelectedIndex == 1) {
-          _controller.setMainSelectedIndex(0);
+        Controller _controller = Provider.of<Controller>(context, listen: false);
+        if (_controller.homeIndex != ArticleTitlesPageIndex) {
+          _controller.jumpToHome(ArticleTitlesPageIndex);
           return false;
         } else
           return true;
       },
-      child: PageView(
-          reverse: true,
-          onPageChanged: (i) {
-            _articleTitles.currentArticleIndex = i;
-            // used to highlight aritcleTitlePage item
-            _controller.setSelectedArticleID(_articleTitles.filterTitles[i].id);
-          },
-          controller: _controller.articlePageController,
-          children: _articleTitles.filterTitles.map((d) {
-            return ArticlePage(d.id);
-          }).toList()),
+      child: Selector<ArticleTitles, List<ArticleTitle>>(
+          selector: (context, articleTitles) => articleTitles.filterTitles,
+          builder: (context, filterTitles, child) {
+            Controller _controller = Provider.of<Controller>(context, listen: false);
+            _controller.articlePageViewController = PageController(initialPage: _controller.articleIndex);
+            return PageView(
+                reverse: true,
+                onPageChanged: (i) =>
+                    _controller.setSelectedArticleID(filterTitles[i].id), // used to highlight aritcleTitlePage item
+                controller: _controller.articlePageViewController,
+                children: filterTitles.map((d) => ArticlePage(d.id)).toList());
+          }),
     );
   }
 }