+++ lib/pages/home.dart
@@ -1,64 +1,44 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import './article_titles.dart';
-import 'article_page_view.dart';
+import './explorer.dart';
+import './article_page_view.dart';
 import '../models/controller.dart';
+import '../themes/base.dart';
 
-class HomePage extends StatefulWidget {
-  @override
-  HomePageState createState() => HomePageState();
-}
-
-class HomePageState extends State<HomePage> {
-  Controller _controller;
-  //ArticleTitles _articleTitles;
-
-  static List<BottomNavigationBarItem> bottomNavigationBarItem =
-      <BottomNavigationBarItem>[
-    BottomNavigationBarItem(
-      icon: Icon(Icons.home),
-      title: Text('Home'),
-    ),
-    BottomNavigationBarItem(
-      icon: Icon(Icons.library_books),
-      title: Text('Article'),
-    ),
-    BottomNavigationBarItem(
-      icon: Icon(Icons.local_library),
-      title: Text('Lib'),
-    ),
-  ];
-  @override
-  void initState() {
-    super.initState();
-    _controller = Provider.of<Controller>(context, listen: false);
+class HomePage extends StatelessWidget {
+  SnackBar getSnackBar(String info) {
+    return SnackBar(
+      backgroundColor: mainColor,
+      content: Text(
+        info,
+        textAlign: TextAlign.center,
+      ),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    // need use Consumer Controller make sure tap new article enter correct article detail
-    return Consumer<Controller>(
-      builder: (context, currentController, child) {
-        return Scaffold(
-          body: PageView(
-            physics: NeverScrollableScrollPhysics(),
-            children: [
-              ArticleTitlesPage(),
-              ArticlePageViewPage(),
-              //Center(child: Text('Developing')),
-            ],
-            controller: _controller.mainPageController,
-            onPageChanged: (index) {},
-          ),
-          /*
-          bottomNavigationBar: BottomNavigationBar(
-            items: bottomNavigationBarItem,
-            currentIndex: currentController.mainSelectedIndex,
-            onTap: _onItemTapped,
-          ),
-          */
-        );
-      },
-    );
+    return Scaffold(
+        body: Selector<Controller, String>(
+            shouldRebuild: (previous, next) => next != "", //even the snackBarInfo not change, still need show
+            selector: (context, controller) => controller.snackBarInfo,
+            builder: (context, snackBarInfo, child) {
+              if (snackBarInfo != null)
+                Future.delayed(Duration.zero, () {
+                  Scaffold.of(context).showSnackBar(getSnackBar(snackBarInfo));
+                  Provider.of<Controller>(context, listen: false).snackBarInfo = null;
+                });
+              return child;
+            },
+            child: PageView(
+                //disable scroll
+                physics: NeverScrollableScrollPhysics(),
+                children: [
+                  ArticleTitlesPage(),
+                  ArticlePageViewPage(),
+                  ExplorerPage(),
+                ],
+                controller: Provider.of<Controller>(context, listen: false).homePageViewController)));
   }
 }