+++ lib/store/store.dart
@@ -1,23 +1,34 @@
 import 'package:dio/dio.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:sqflite/sqflite.dart';
 
 class Store {
   static const baseURL = "https://english.bigzhu.net/api/";
-}
+  static const PATH = "assets/db/wordwise.db";
+  //static const baseURL = "http://192.168.43.231:3004/api/";
+  static SharedPreferences _prefs;
 
-Dio getDio() {
-  Dio dio = new Dio();
-  // 发送请求前加入 token
-  dio.interceptors.add(InterceptorsWrapper(onRequest: (Options options) async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    String accessTokenShare = prefs.getString('accessToken');
-    options.headers["token"] = accessTokenShare;
-    return options; //continue
-  }, onError: (DioError e) {
-    // Do something with response error
-    print(e.toString());
-    throw e;
-  }));
+  static Database database;
+  static Map wordwiseMap = Map<String, String>();
+  static Map noWordwiseMap = Map<String, String>();
 
-  return dio;
+  static Dio dio() {
+    Dio _dio = Dio();
+    // 发送请求前加入 token
+    _dio.interceptors.add(InterceptorsWrapper(onRequest: (Options options) async {
+      if (_prefs == null) await Store.prefs;
+      String accessTokenShare = _prefs.getString('accessToken');
+      options.headers["token"] = accessTokenShare;
+      return options; //continue
+    }, onError: (DioError e) {
+      throw e;
+    }));
+    return _dio;
+  }
+
+  static Future<SharedPreferences> get prefs async {
+    if (_prefs != null) return _prefs;
+    _prefs = await SharedPreferences.getInstance();
+    return _prefs;
+  }
 }